Test the basic API set up
When testing, I ran into a few issues to do with nesting. I decided to
be explicit about what nested attributes to include in the recipes
controller as this forces me to really think about the data, and what
should be exposed where.

Also had some issues with formatting matching my tests, so am using
the ActiveModelSerializers::SerializableResource but specifying which
serializer to use.

Also decided to use the json_api adapter as this seems to be
standard, and makes the formatting consistent.